==> 2_machine_learning/3_Regresja logistyczna/rozwiazania/2_interpretacja_modelu.ipynb <==
# Interpretacja modelu

# Import wymaganych bibliotek
from  sklearn.linear_model import LogisticRegression
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns

os.chdir('../')

# puść ten kod, 
# jeżeli wywołujesz plik  w folderze rozwiąznaia, 
# a ramka danych znajduje się w folderze data
import os 
os.chdir('../')

# Pobranie danych

df = pd.read_csv('data/Cancer_Data.csv')
train_x = pd.read_csv('data/cancer_train_x.csv').iloc[:,1:]
test_x = pd.read_csv('data/cancer_test_x.csv').iloc[:,1:]
train_y = pd.read_csv('data/cancer_train_y.csv')['target']
test_y = pd.read_csv('data/cancer_test_y.csv')['target']

## Estymacja modelu
Regularyzacja jest metodą ograniczającą przeuczenie. Nakładana jest kara na funkcję celu, która minimalizuje nadmierne dopasowanie. 
W modelu regresji logistycznej mamy regularyzację wg parametrów l1 oraz l2. Siła regularyzacji ustalana jest parametrem C, gdzie niższe wartości, dają silniejszą regularyzację. 

Więcej informacji można znaleźć tu: https://medium.com/@rithpansanga/logistic-regression-and-regularization-avoiding-overfitting-and-improving-generalization-e9afdcddd09d

## dodanie zmiennej kategorycznej
train_x['perimeter_se_kat'] = (train_x['perimeter_se'] >=df['perimeter_se'].mean()).astype(int)
test_x['perimeter_se_kat'] = (test_x['perimeter_se'] >=df['perimeter_se'].mean()).astype(int)

# Usunięcie wersji nominalnej zmiennej
del train_x['perimeter_se']
del test_x['perimeter_se']

# Stworzenie obiektu modelu
model_1 = LogisticRegression(C=0.4)

# Estymacja 
model_1.fit(train_x, train_y)

#Współczynniki modelu
model_1.coef_

# Nazwy zmiennych
model_1.feature_names_in_

# Statystyki zmiennych x
train_x.describe()

Nie możemy porównywać pomiędzy sobą współczynników, ponieważ zakres zmiennych jest różny. Do porównania należałoby wystandaryzować zmienne

# Interpretacja zmiennej perimeter_worst
var = 'perimeter_worst'
index = list(model_1.feature_names_in_).index(var)
OR = np.round(np.exp(model_1.coef_[0][index])-1,3)
print(f'Wraz ze wzrostem {var} o jednostkę, szansa zajścia zdarzenia wzrasta o {np.round(OR*100,1)}%')

# Interpretacja zmiennej kategorycznej
var = 'perimeter_se_kat'
index = list(model_1.feature_names_in_).index(var)
OR = np.round(1-np.exp(model_1.coef_[0][index]),3)
print(f'W grupie o wartości zmiennej {var}=1,czyli w grupie powyżej średniej, szansa zajścia zdarzenia jest niższa o {np.round(OR*100,1)}% niż w grupie poniżej średniej.')

## Interacje i przekształcenia zmiennych
Model regresji logistycznej wykrywa jedynie zależności liniowe, dlatego warto testować:
1. przekształcenia zmiennych (np. logarytm, potęga, 1/x,...)
2. Interakcje - iloczyn dwóch zmiennych (mogą też być kategoryczne)


# Iloczyn zmiennej kategorycznej z nominalną
train_x['perim_il'] = train_x['perimeter_se_kat'] * train_x['perimeter_worst']
test_x['perim_il'] = test_x['perimeter_se_kat'] * test_x['perimeter_worst']

# Stworzenie obiektu modelu
model_1 = LogisticRegression(C=0.4)

# Estymacja 
model_1.fit(train_x,train_y)

#Współczynniki modelu
model_1.coef_

# Nazwy zmiennych
model_1.feature_names_in_

# Interpretacja interakcji
var = 'perim_il'
index = list(model_1.feature_names_in_).index(var)
OR = np.round(1-np.exp(model_1.coef_[0][index]),3)
print(f'W grupie o wartości zmiennej {var}=1 czyli w grupie powyżej średniej wraz ze wzrostem wartości perimeter_worst o jednostke szansa zajścia zdarzenia jest niższa dodatkowo {OR*100}%.')

Interpretacja przekształcenia zmiennej odbyba się w taki sam sposób jak podstawowyc zmiennych, zmienia się tylko definicja zmiennej.
